issue1 処理の一部を別のファイルに移動
