+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_study/count_up_screen.dart';
 
 void main() {
   runApp(const MyApp());
@@ -13,65 +14,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Demo',
-      home: CountUpApp(
-        title: title,
-      ),
-    );
-  }
-}
-
-class CountUpApp extends StatefulWidget {
-  const CountUpApp({super.key, required this.title});
-
-  final String title;
-
-  @override
-  CountUpState createState() => CountUpState();
-}
-
-class CountUpState extends State<CountUpApp> {
-  int _counter = 0;
-
-  void _incrementCounter() {
-    setState(() {
-      if (_counter < 100) {
-        _counter++;
-      } else {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('100が限界です'),
-          ),
-        );
-      }
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              '$_counter',
-              style: const TextStyle(fontSize: 48.0),
-            ),
-            const SizedBox(height: 20),
-            ElevatedButton.icon(
-              onPressed: _incrementCounter,
-              icon: const Icon(
-                Icons.add_circle,
-                color: Colors.white,
-              ),
-              label: const Text('Count Up!'),
-            ),
-          ],
-        ),
-      ),
+      home: CountUpApp(title: title)
     );
   }
 }
